Move extension lists to individual files
