+++ src/zepben/ewb/boilerplate/relations/battery_control_list.py
@@ -1,0 +1,24 @@
+#  Copyright 2026 Zeppelin Bend Pty Ltd
+#  This Source Code Form is subject to the terms of the Mozilla Public
+#  License, v. 2.0. If a copy of the MPL was not distributed with this
+#  file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+from __future__ import annotations
+
+from zepben.ewb import BatteryControlMode
+from zepben.ewb.boilerplate.collections.mrid_list import LazyMridList
+
+
+class BatteryControlList(LazyMridList):
+    def get_by_mode(self, control_mode: BatteryControlMode):
+        """
+        Get the `BatteryControl` identified by its `control_mode`
+
+        `control_mode` the `BatteryControlMode` of the desired `BatteryControl`
+        Returns The `BatteryControl` with the specified `control_mode` if it exists
+        Raises `KeyError` if a `BatteryControl` with `control_mode` wasn't present.
+        """
+        for control in self:
+            if control.control_mode == control_mode:
+                return control
+        raise IndexError(f"No BatteryControl with a control_mode of {control_mode} was found in BatteryUnit {str(self)}")
