+++ src/zepben/ewb/model/cim/iec61970/base/core/conducting_equipment.py
@@ -14,24 +14,16 @@ from dataclasses import field
 
 from typing_extensions import deprecated
 
-from zepben.ewb.boilerplate.collections.mrid_list import LazyMridList
 from zepben.ewb.boilerplate.backfill import Backfill
 from zepben.ewb.model.cim.iec61970.base.core.terminal import Terminal
 from zepben.ewb.model.cim.iec61970.base.core.equipment import Equipment
+from zepben.ewb.boilerplate.relations.terminal_list import TerminalsList
 from zepben.ewb.util import require
 from zepben.ewb.boilerplate.dataclass_base import zb_dataclass
 
 if TYPE_CHECKING:
     from zepben.ewb.model.cim.iec61970.base.core.base_voltage import BaseVoltage
 
-
-class TerminalsList(LazyMridList[Terminal]):
-
-    def get_by_sequence_number(self, sequence_number: int) -> Terminal:
-        term = next((it for it in self if it.sequence_number == sequence_number), None)
-        if term is None:
-            raise IndexError(f"No Terminal with sequence_number {sequence_number} was found in ConductingEquipment {str(self.instance)}")
-        return term
 
 @zb_dataclass
 class ConductingEquipment(Equipment, metaclass=ABCMeta):
