+++ src/zepben/ewb/model/cim/iec61970/base/core/curve.py
@@ -12,38 +12,11 @@ from abc import ABCMeta
 from typing_extensions import deprecated
 
 from zepben.ewb import zb_dataclass
-from zepben.ewb.boilerplate.collections.lazy_collection import LazyCollection
 from zepben.ewb.model.cim.iec61970.base.core.curve_data import CurveData
 from zepben.ewb.model.cim.iec61970.base.core.identified_object import IdentifiedObject
+from zepben.ewb.boilerplate.relations.curve_data_list import CurveDataList
 from zepben.ewb.util import require
 
-
-class CurveDataList(LazyCollection[CurveData]):
-
-    def get(self, x: float) -> CurveData:
-        """
-        Get the :class:`CurveData` identified by its `x_value`.
-
-        :param x: The X value of the required :class:`CurveData`.
-        :returns: The :class:`CurveData` with the specified `x` if it exists.
-        :raises KeyError: When no `CurveData` was found with `x`.
-        """
-        curve_data = next((it for it in self if it.x_value == x), None)
-        if curve_data:
-            return curve_data
-        raise KeyError(x)
-
-    def remove_data_at(self, x: float) -> CurveData:
-        """
-        Disassociate a :class:`CurveData` from this collection based on its `x_value`.
-
-        :param x: The :class:`CurveData` to disassociate from this :class:`Curve`.
-        :returns: A reference to the removed :class:`CurveData`.
-        :raises IndexError: If no :class:`CurveData` with a value of `x` was not associated with this :class:`Curve`.
-        """
-        data = self.get(x)
-        self.remove(data)
-        return data
 
 @zb_dataclass
 class Curve(IdentifiedObject, metaclass=ABCMeta):
