+++ src/zepben/ewb/model/cim/iec61970/base/wires/per_length_phase_impedance.py
@@ -13,35 +13,11 @@ from typing import List, Generator
 from typing_extensions import deprecated
 
 from zepben.ewb import zb_dataclass
-from zepben.ewb.boilerplate.collections.lazy_collection import LazyCollection
 from zepben.ewb.model.cim.iec61970.base.wires.per_length_impedance import PerLengthImpedance
 from zepben.ewb.model.cim.iec61970.base.wires.phase_impedance_data import PhaseImpedanceData
 from zepben.ewb.model.cim.iec61970.base.wires.single_phase_kind import SinglePhaseKind
+from zepben.ewb.boilerplate.relations.phase_impedance_data_list import PhaseImpedanceDataList
 from zepben.ewb.util import require, none
-
-
-class PhaseImpedanceDataList(LazyCollection):
-    def get(self, from_phase: SinglePhaseKind, to_phase: SinglePhaseKind) -> PhaseImpedanceData:
-        """
-        Get the matrix entry for the corresponding to and from phases.
-
-        :param from_phase: The from_phase to lookup.
-        :param to_phase: The to_phase to lookup.
-        :returns: The :class:`PhaseImpedanceData` with the specified `from_phase` and `to_phase` if it exists.
-        :raises KeyError: When no `PhaseImpedanceData` was found with a matching `from_phase` and `to_phase`.
-        """
-        phase_impedance_data = next((it for it in self if it.from_phase == from_phase and it.to_phase == to_phase), None)
-        if phase_impedance_data:
-            return phase_impedance_data
-
-        raise KeyError((from_phase, to_phase))
-
-    @property
-    def diagonal(self):
-        """
-        Get only the diagonal elements of the matrix, i.e toPhase == fromPhase.
-        """
-        return (pid for pid in self if pid.from_phase == pid.to_phase)
 
 
 @zb_dataclass
